Paper over thread ID truncation on 64-bit systems
Truncating 64-bit thread IDs to 32-bits can result in collisions on
systems with large numbers of CPUs. GetCurrentThreadId() does this, and
there are fairly easy to reproduce collisions on systems with ~128 CPU
cores.

Rather than threading a new type everywhere to resolve this, this patch
simply XORs in the high-order bits of the pthread_t to increase the
entropy in thread identifiers.

This fixes failures in TestSynchCritical.